use git tag manually if latest is version specified
